fix some endpoints

+++ src/test/java/services/GameService.java
@@ -1,9 +1,9 @@
 package services;
 
 import assections.AssertableResponse;
+import db_models.game.Game;
 import helpers.Utils;
 import db_models.game.DLC;
-import models.game.Game;
 import models.game.UpdField;
 
 import java.util.List;
@@ -18,11 +18,11 @@ public class GameService extends WebService {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt).get("games").then());
     }
 
-    public AssertableResponse getGame(int id) {
+    public AssertableResponse getGame(long id) {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt).get("games/" + id).then());
     }
 
-    public AssertableResponse deleteGame(int id) {
+    public AssertableResponse deleteGame(long id) {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt).delete("games/" + id).then());
     }
 
@@ -31,12 +31,12 @@ public class GameService extends WebService {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt).body(game).post("games").then());
     }
 
-    public AssertableResponse deleteListDlc(int id, List<DLC> dlcList) {
+    public AssertableResponse deleteListDlc(long id, List<DLC> dlcList) {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt)
                 .body(dlcList).delete(String.format("games/%d/dlc", id)).then());
     }
 
-    public AssertableResponse updateListDlc(int id, List<DLC> dlcList) {
+    public AssertableResponse updateListDlc(long id, List<DLC> dlcList) {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt).body(dlcList).put("games/" + id).then());
     }
 
@@ -44,7 +44,7 @@ public class GameService extends WebService {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt).body(game).post("games").then());
     }
 
-    public AssertableResponse updateField(int id, UpdField updField) {
+    public AssertableResponse updateField(long id, UpdField updField) {
         return new AssertableResponse(requestSpec.auth().oauth2(jwt).body(updField).put(String.format("games/%d/updateField", id)).then());
     }
 }
